+++ NEWS.md
@@ -1,8 +1,9 @@
 # plm 2.2-4 (in development)
 * exported: fixef.pggls and deprecated methods/function formula.dynformula,
-  print.dynformula, pvcov: exports were lost in plm 2.0-0, exported again.
+  print.dynformula, pvcovHC: exports were lost in plm 2.0-0, exported again.
 * pwaldtest.pgmm: non-exported function wald() now exported as method
   pwaldtest.pgmm.
+* mtest, sargan, pwaldtest: added for each string for alternative hypothesis.
 
 # plm 2.2-3
 * IGNORE_RDIFF_BEGIN/END added on tests and man pages